requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-redis==5.4.0
idna==3.10
JSON-log-formatter==1.1
multidict==6.1.0
propcache==0.3.0
psycogreen==1.0.2
psycopg2-binary==2.9.10
pytz==2025.1
redis==5.2.1
requests==2.32.3
setuptools==75.8.1
sqlparse==0.5.3
urllib3==2.3.0
wheel==0.45.1
yarl==1.18.3
ccxt==4.4.75
eventlet==0.39.1
protobuf==5.29.3
grpcio==1.70.0
grpcio-tools==1.70.0
ntplib==0.4.0
python-dateutil==2.9.0.post0